move openshift api functions from shared functions to resources directory
