+++ Makefile
@@ -15,6 +15,7 @@ OPERATOR_SDK_VERSION=0.15.1
 AUTH_TOKEN=$(shell curl -sH "Content-Type: application/json" -XPOST https://quay.io/cnr/api/v1/users/login -d '{"user": {"username": "$(QUAY_USERNAME)", "password": "${QUAY_PASSWORD}"}}' | jq -r '.token')
 TEMPLATE_PATH="$(shell pwd)/templates/monitoring"
 INTEGREATLY_OPERATOR_IMAGE ?= $(REG)/$(ORG)/$(PROJECT):v$(TAG)
+IDP_PASSWORD=Password1
 
 # If openapi-gen is available on the path, use that; otherwise use it through
 # "go run" (slower)
@@ -60,6 +61,11 @@ setup/service_account:
 .PHONY: setup/git/hooks
 setup/git/hooks:
 	git config core.hooksPath .githooks
+
+
+.PHONY: setup/testing_idp
+setup/testing_idp:
+	PASSWORD=$(IDP_PASSWORD) ./scripts/setup-sso-idp.sh
 
 .PHONY: code/run
 code/run: code/gen cluster/prepare/smtp cluster/prepare/dms cluster/prepare/pagerduty
@@ -145,9 +151,8 @@ test/e2e: cluster/cleanup cluster/cleanup/crds cluster/prepare cluster/prepare/c
 	$(OPERATOR_SDK) --verbose test local ./test/e2e --namespace="$(NAMESPACE)" --go-test-flags "-timeout=60m" --debug --image=$(INTEGREATLY_OPERATOR_IMAGE)
 
 .PHONY: test/functional
-test/functional:
+test/functional: setup/testing_idp
 	# Run the functional tests against an existing cluster. Make sure you have logged in to the cluster.
-	PASSWORD=Password1 ./scripts/setup-sso-idp.sh
 	go clean -testcache && go test ./test/functional
 
 .PHONY: install/olm
